chore:rename build-loader
